add ReadOnly component
