+++ src/modals/ConnectAccounts/Wrappers.ts
@@ -317,71 +317,3 @@ export const Separator = styled.div`
   opacity: 0.1;
   margin: 1.5rem 0rem;
 `;
-
-export const ReadOnlyWrapper = styled.div`
-  box-sizing: border-box;
-  border-radius: 0.75rem;
-  width: 100%;
-  margin: 1rem 0;
-  border-radius: 0.5rem;
-  font-size: 1rem;
-  background: ${buttonPrimaryBackground};
-  transition: background 0.15s;
-  color: ${textPrimary};
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: flex-start;
-  min-height: 3.5rem;
-
-  > .content {
-    box-sizing: border-box;
-    padding: 0 1rem;
-    width: 100%;
-  }
-
-  .accounts {
-    margin-top: 1rem;
-    width: 100%;
-  }
-
-  .account {
-    box-sizing: border-box;
-    width: 100%;
-    border: 1px solid ${borderPrimary};
-    border-radius: 0.75rem;
-    margin: 1rem 0;
-    padding: 1rem;
-    display: flex;
-    flex-flow: row wrap;
-    transition: border 0.1s;
-
-    > div {
-      color: ${textSecondary};
-      transition: opacity 0.2s;
-
-      &:first-child {
-        box-sizing: border-box;
-        flex: 1;
-        display: flex;
-        flex-flow: row wrap;
-        overflow: hidden;
-        text-overflow: ellipsis;
-        white-space: nowrap;
-      }
-      &:last-child {
-        padding-left: 2rem;
-        opacity: 0.25;
-      }
-    }
-
-    &:hover {
-      > div:last-child {
-        opacity: 1;
-      }
-    }
-
-    &:hover {
-      border-color: ${borderSecondary};
-    }
-  }
-`;
